+++ CHANGELOG.md
@@ -7,8 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [[5.0.13](https://github.com/multiversx/mx-sdk-dapp/pull/1548)] - 2025-08-06
 
-## [[5.0.12](https://github.com/multiversx/mx-sdk-dapp/pull/1541)] - 2025-08-01
+- [Fix provider restoration when async storage is setup in the client](https://github.com/multiversx/mx-sdk-dapp/pull/1547)
+
+## [[5.0.12](https://github.com/multiversx/mx-sdk-dapp/pull/1541)] - 2025-08-06
 
 - [Fix WalletConnect Provider issues related to consecutive signings](https://github.com/multiversx/mx-sdk-dapp/pull/1546)
 - [Removed loginInfo unused members](http://github.com/multiversx/mx-sdk-dapp/pull/1543)